feat: date 상태 관리 및 파일 이동
