+++ src/routes/Dashboard/DateInput/datepicker.scss
@@ -108,3 +108,39 @@
     background: colors.$CALENDER_RANGE_BG;
   }
 }
+
+.calender-header {
+  width: 100%;
+  height: 60px;
+  line-height: 60px;
+  color: colors.$TEXT_WHITE;
+  text-align: center;
+  background: colors.$PRIMARY;
+  border-top-left-radius: 10px;
+  border-top-right-radius: 10px;
+}
+
+.calender-main {
+  position: relative;
+  display: flex;
+  width: 100%;
+  border-bottom: 1px solid colors.$BORDER;
+
+  .monthly-header {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0 10px;
+
+    .prev,
+    .next {
+      display: inline-flex;
+      align-items: center;
+      justify-content: center;
+      width: 30px;
+      height: 30px;
+      background: colors.$BG_G;
+      border-radius: 50%;
+    }
+  }
+}
