+++ src/routes/Dashboard/dashboard.module.scss
@@ -11,46 +11,5 @@
       font-size: 26px;
       font-weight: 900;
     }
-
-    .dateWrapper {
-      position: relative;
-      width: 100%;
-    }
-
-    .calenderHeader {
-      width: 100%;
-      height: 60px;
-      line-height: 60px;
-      color: colors.$TEXT_WHITE;
-      text-align: center;
-      background: colors.$PRIMARY;
-      border-top-left-radius: 10px;
-      border-top-right-radius: 10px;
-    }
-
-    .calenderMain {
-      position: relative;
-      display: flex;
-      width: 100%;
-      border-bottom: 1px solid colors.$BORDER;
-
-      .monthlyHeader {
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-        padding: 0 10px;
-
-        .prevButton,
-        .nextButton {
-          display: inline-flex;
-          align-items: center;
-          justify-content: center;
-          width: 30px;
-          height: 30px;
-          background: colors.$BG_G;
-          border-radius: 50%;
-        }
-      }
-    }
   }
 }
