+++ src/routes/Dashboard/states.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil'
+
+export const dateRangeState = atom<string[]>({
+  key: '#dateRangeState',
+  default: ['2022-04-14', '2022-04-20'],
+})
